+++ lib/td_dd_web/resolvers/structures.ex
@@ -8,4 +8,14 @@ defmodule TdDdWeb.Resolvers.Structures do
   def data_structure_versions(_parent, args, _resolution) do
     {:ok, DataStructures.list_data_structure_versions(args)}
   end
+
+  def domain_id(%{domain_ids: domain_ids}, _args, _resolution) do
+    domain_id =
+      case domain_ids do
+        [domain_id | _] -> domain_id
+        _ -> nil
+      end
+
+    {:ok, domain_id}
+  end
 end