+++ lib/td_dd_web/swagger_definitions.ex
@@ -377,7 +377,7 @@ defmodule TdDdWeb.SwaggerDefinitions do
             system_id(:integer, "System Id", required: true)
             group(:string, "Data Structure group")
             name(:string, "Data Structure name")
-            domain_id([:integer, :null], "Domain Id")
+            domain_ids(:array, "Domain Ids")
             class([:string, :null], "Data Structure class")
             description([:string, :null], "Data Structure description")
             external_id([:string, :null], "Data Structure external id")
@@ -403,6 +403,7 @@ defmodule TdDdWeb.SwaggerDefinitions do
             },
             group: "Data Structure group",
             name: "Data Structure name",
+            domain_ids: [1, 2],
             description: "Data Structure description",
             type: "Csv",
             confidential: "Data Structure confidentiality",