+++ priv/repo/migrations/20220217163931_alter_rule_implementations_timestamps.exs
@@ -1,12 +1,10 @@
 defmodule TdDd.Repo.Migrations.AlterRuleImplementationsTimestamps do
   use Ecto.Migration
 
-  use Ecto.Migration
-
   def change do
     alter table("rule_implementations") do
-      modify(:inserted_at, :utc_datetime)
-      modify(:updated_at, :utc_datetime)
+      modify(:inserted_at, :utc_datetime, from: :naive_datetime)
+      modify(:updated_at, :utc_datetime, from: :naive_datetime)
     end
   end
 end